Add assertions to InputParser class
The Mark, Unmark and Delete commands require a task number to be
entered. The task number needs to exist in the current Task List.

Adding assertions to assert that the task number is a valid task number
ensures that we do not run into IndexOutOfBounds error when accessing
the task from the task list.